+++ Package/remp/js/components/FormValidator.vue
@@ -15,6 +15,7 @@ export default {
 
         $(this.$el).closest('form').on('submit', function () {
             var form = this;
+            console.log('submitujem tadialto');
 
             if ($(form).attr('data-valid')) {
                 return true;