+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added the configurable option to prioritize banners on same position. To enable set `env` variable `PRIORITIZE_BANNERS_ON_SAME_POSITION`. remp/remp#1167
 - Added information about where to display to the detail page of campaign. remp/remp#1009
 - Changed the form of editing banner's custom JS - now JS code is displayed formatted and linted in separate window. remp/remp#1171
+- Added support for the inline editing of variables in the custom JS of banner. remp/remp#1172
 
 ### [Mailer]